moved inline_thresh test to test_openers & shared fixtures to conftest
